Alphabetize the wordlist
Solely for convenience, and because I see no reason to not have it
alphabetized.

If anything, we could try to protect from someone shoulder-surfing and
noticing some of the digits; but then we should randomise the list---
which I doubt it was prior to this commit. (Yeah, "NINE NINE NINE ...")
I'd suggest we don't go there.